Refactor config parsing and add RTSP test
